No if_state_lock in ifid fast-path
As the actual state of the interface will not be changed, the "global"
if_state_lock does not appear to be required here. The actual timestamp
update is still atomic (internal lock per if state object).
This makes sure the fast-path will not be blocked by a concurrent
timeout-check or a "slow-path" IFID processing.