Remove capnp dependency from scion-pki WIP
capnp comes in over a bunch of different ways:

  bazel query "allpaths(//go/scion-pki:go_default_library, @com_zombiezen_go_capnproto2//...)" --output graph

The infra library only has dead code related to this.

The rest is related to handling revocations / sending revocations to sciond. For testing, removed this code (which is needed though).
It seems that it should be possible to redefine the revocation format so they can be parsed without capnp.

Btw. the binary size did not decrease as significantly as expected after this.