WIP doc: SPAO replace hash-then-mac scheme, no SHA-1
Replace the hash-then-mac SHA1-AES-CBC algorithm type. The main reason
for this change is that SHA1 has known weaknesses, and using it looks
bad, regardless of whether or not these weaknesses could possibly be
exploited in this case.

Instead of just replacing SHA-1 with a different hash function like
SHA-2, SHA-3 or BLAKE, we use a MAC, specifically AES-CMAC, for this
purpose. This is feasible because we have a shared key anyway.

The advantage of this is that

................ reeeeeewiiiiind

Discussed with Marc Frei

- there is not really a good reason to include the first MAC in the
  second MAC.
  Could be simplified!
- But... we should double check that it's not actually much slower than
  the hash thing
- Plus, the assumption that the receiving host has the shared key might
  not be valid.